convert telescope to use ev instead of libevent
libevent is a very cool library, I like the APIs and enjoy using
it.  However, telescope is not as large and doesn't have as many
file descriptor, so libevent is quite ``too big'' for our needs.

ev started as a small event loop on top of poll(2) for amused, and
can be used here too, it just needed to grow the ability to handle
several timers, as we need quite some on telescope (in fact, we use
more timers than fds!)

+++ control.h
@@ -19,6 +19,6 @@
 #ifndef	SMALL
 int	control_init(char *);
 int	control_listen(int);
-void	control_accept(int, short, void *);
-void	control_dispatch_imsg(int, short, void *);
+void	control_accept(int, int, void *);
+void	control_dispatch_imsg(int, int, void *);
 #endif	/* SMALL */
